alx_travel_app/requirements.txt
Django>=4.0,<5.0
djangorestframework
django-cors-headers
drf-yasg
mysqlclient
django-environ
celery
